conf mtrx

+++ supervised_learning/error_analysis/0-create_confusion.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python3
 """ Confusing Matrix""" 
-
 import numpy as np
 
 
@@ -9,7 +8,7 @@ def create_confusion_matrix(labels, logits):
     true_classes = np.argmax(labels, axis=1)
     predicted_classes = np.argmax(logits, axis=1)
     classes = labels.shape[1]
-    confusion_matrix = np.zeros((classes, classes), dtype=int)
+    confusion_matrix = np.zeros((classes, classes))
     # populate confusion matrix
     for true, pred in zip(true_classes, predicted_classes):
         confusion_matrix[true][pred] += 1
